backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
pydantic-settings
Pillow
mysql-connector-python
email-validator
python-multipart
alembic
